requirements.txt
streamlit==1.35.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.3
seaborn==0.13.1
plotly==5.18.0
supabase==2.3.0
python-dotenv==1.0.0
pillow==10.2.0 